Move cfsites stuff in a dedicated module
